requirements.txt
absl-py==0.11.0
aiofiles==0.6.0
alembic==1.5.8
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.3+cu101
entrypoints==0.3
filterpy==1.4.5
future==0.18.2
fvcore==0.1.2.post20201218
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
imageio==2.9.0
iopath==0.1.2
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.3
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.8
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
portalocker==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.1
Pygments==2.7.3
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyWavelets==1.1.1
pywinpty==0.5.7
PyYAML==5.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
sanic==20.12.2
scikit-image==0.17.2
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.23
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2021.2.1
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.2
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
